feat: написана функция запроса на ревью
